backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1
google-genai==1.0.0
httpx==0.28.1
python-multipart==0.0.20
